requirements.txt
discord.py==2.3.2
requests~=2.27.1
packaging~=21.3
python-dotenv~=0.20.0
readerwriterlock~=1.0.9
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
starlette==0.27.0
PyYAML~=6.0.1